requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1

# AI/ML Libraries
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Time Series Analysis
statsmodels==0.14.1
prophet==1.1.5

# Data Processing
scipy==1.11.4
openpyxl==3.1.2
xlrd==2.0.1

# Database
pymongo==4.6.0
psycopg2-binary==2.9.9  # PostgreSQL driver
PyMySQL==1.1.0          # MySQL driver
pyodbc==5.0.1           # SQL Server driver

# File Processing
pypdf2==3.0.1
reportlab==4.0.7

# Utilities
python-dotenv==1.0.0
pydantic==2.5.1
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0 